BCV-3 add package 'server' to the root of server module
