+++ server/build.gradle
@@ -14,7 +14,7 @@ apply plugin: 'org.springframework.boot'
 apply plugin: 'io.spring.dependency-management'
 
 apply plugin: "application"
-mainClassName = 'dsx.bcv.app.Application'
+mainClassName = 'dsx.bcv.server.app.Application'
 
 bootJar {
 }
