--- server/src/main/java/dsx/bcv/app/Application.java
@@ -1,20 +1,0 @@
-package dsx.bcv.app;
-
-import org.springframework.boot.SpringApplication;
-import org.springframework.boot.autoconfigure.SpringBootApplication;
-import org.springframework.context.annotation.ComponentScan;
-
-@ComponentScan(value = "dsx/bcv")
-@SpringBootApplication
-public class Application {
-
-    public static void main(String[] args) {
-        init();
-        SpringApplication.run(Application.class, args);
-    }
-
-    private static void init(){
-        System.setProperty("server.servlet.context-path", "/bcv");
-        System.getProperties().put("server.port", 9999);
-    }
-}
