--- server/src/main/java/dsx/bcv/exceptions/NotFoundException.java
@@ -1,8 +1,0 @@
-package dsx.bcv.exceptions;
-
-import org.springframework.http.HttpStatus;
-import org.springframework.web.bind.annotation.ResponseStatus;
-
-@ResponseStatus(value = HttpStatus.NOT_FOUND)
-public class NotFoundException extends RuntimeException {
-}
