+++ server/src/main/java/dsx/bcv/server/exceptions/NotFoundException.java
@@ -1,0 +1,8 @@
+package dsx.bcv.server.exceptions;
+
+import org.springframework.http.HttpStatus;
+import org.springframework.web.bind.annotation.ResponseStatus;
+
+@ResponseStatus(value = HttpStatus.NOT_FOUND)
+public class NotFoundException extends RuntimeException {
+}
