+++ server/src/main/java/dsx/bcv/server/services/parsers/TradeField.java
@@ -1,0 +1,14 @@
+package dsx.bcv.server.services.parsers;
+
+public enum TradeField {
+    DateTime,
+    Instrument,
+    TradeType,
+    TradedQuantity,
+    TradedQuantityCurrency,
+    TradedPrice,
+    TradedPriceCurrency,
+    Commission,
+    CommissionCurrency,
+    TradeValueId
+}
