+++ server/src/main/java/dsx/bcv/server/services/parsers/TransactionField.java
@@ -1,0 +1,11 @@
+package dsx.bcv.server.services.parsers;
+
+public enum TransactionField {
+    DateTime,
+    TransactionType,
+    Currency,
+    Amount,
+    Commission,
+    TransactionStatus,
+    TransactionValueId
+}
