--- server/src/main/java/dsx/bcv/services/parsers/TradeField.java
@@ -1,14 +1,0 @@
-package dsx.bcv.services.parsers;
-
-public enum TradeField {
-    DateTime,
-    Instrument,
-    TradeType,
-    TradedQuantity,
-    TradedQuantityCurrency,
-    TradedPrice,
-    TradedPriceCurrency,
-    Commission,
-    CommissionCurrency,
-    TradeValueId
-}
