--- server/src/main/java/dsx/bcv/services/parsers/TransactionField.java
@@ -1,11 +1,0 @@
-package dsx.bcv.services.parsers;
-
-public enum TransactionField {
-    DateTime,
-    TransactionType,
-    Currency,
-    Amount,
-    Commission,
-    TransactionStatus,
-    TransactionValueId
-}
